+++ packages/avif/meta.ts
@@ -10,22 +10,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
- import { EncodeOptions, AVIFTune } from './codec/enc/avif_enc';
+import { EncodeOptions, AVIFTune } from './codec/enc/avif_enc.js';
 
- export { EncodeOptions, AVIFTune };
- 
- export const label = 'AVIF';
- export const mimeType = 'image/avif';
- export const extension = 'avif';
- export const defaultOptions: EncodeOptions = {
-   cqLevel: 33,
-   cqAlphaLevel: -1,
-   denoiseLevel: 0,
-   tileColsLog2: 0,
-   tileRowsLog2: 0,
-   speed: 6,
-   subsample: 1,
-   chromaDeltaQ: false,
-   sharpness: 0,
-   tune: AVIFTune.auto,
- };
+export { EncodeOptions, AVIFTune };
+
+export const label = 'AVIF';
+export const mimeType = 'image/avif';
+export const extension = 'avif';
+export const defaultOptions: EncodeOptions = {
+  cqLevel: 33,
+  cqAlphaLevel: -1,
+  denoiseLevel: 0,
+  tileColsLog2: 0,
+  tileRowsLog2: 0,
+  speed: 6,
+  subsample: 1,
+  chromaDeltaQ: false,
+  sharpness: 0,
+  tune: AVIFTune.auto,
+};