+++ packages/jxl/utils.ts
@@ -24,17 +24,20 @@ export function initEmscriptenModule<T extends EmscriptenWasm.Module>(
   let instantiateWasm;
 
   if (wasmModule) {
-    instantiateWasm = (imports: WebAssembly.Imports, callback: (instance: WebAssembly.Instance) => void) => {
+    instantiateWasm = (
+      imports: WebAssembly.Imports,
+      callback: (instance: WebAssembly.Instance) => void,
+    ) => {
       const instance = new WebAssembly.Instance(wasmModule, imports);
       callback(instance);
       return instance.exports;
-    }
+    };
   }
 
   return moduleFactory({
     // Just to be safe, don't automatically invoke any wasm functions
     noInitialRun: true,
     instantiateWasm,
-    ...moduleOptionOverrides
+    ...moduleOptionOverrides,
   });
 }