+++ packages/oxipng/optimise.ts
@@ -16,8 +16,8 @@
  * - Use mutli-threading only in a Worker context
  * - Use multi-threading only if the browser has hardware concurrency > 1
  */
-import type { InitInput } from './codec/pkg/squoosh_oxipng';
-import { defaultOptions, OptimiseOptions } from './meta';
+import type { InitInput } from './codec/pkg/squoosh_oxipng.js';
+import { defaultOptions, OptimiseOptions } from './meta.js';
 import { threads } from 'wasm-feature-detect';
 
 async function initMT(moduleOrPath?: InitInput) {
@@ -25,7 +25,7 @@ async function initMT(moduleOrPath?: InitInput) {
     default: init,
     initThreadPool,
     optimise,
-  } = await import('./codec/pkg-parallel/squoosh_oxipng');
+  } = await import('./codec/pkg-parallel/squoosh_oxipng.js');
   await init(moduleOrPath);
   await initThreadPool(globalThis.navigator.hardwareConcurrency);
   return optimise;
@@ -33,7 +33,7 @@ async function initMT(moduleOrPath?: InitInput) {
 
 async function initST(moduleOrPath?: InitInput) {
   const { default: init, optimise } = await import(
-    './codec/pkg/squoosh_oxipng'
+    './codec/pkg/squoosh_oxipng.js'
   );
   await init(moduleOrPath);
   return optimise;
@@ -41,14 +41,20 @@ async function initST(moduleOrPath?: InitInput) {
 
 let wasmReady: ReturnType<typeof initMT | typeof initST>;
 
-export async function init(moduleOrPath?: InitInput): Promise<ReturnType<typeof initMT | typeof initST>> {
+export async function init(
+  moduleOrPath?: InitInput,
+): Promise<ReturnType<typeof initMT | typeof initST>> {
   if (!wasmReady) {
-    const hasHardwareConcurrency = globalThis.navigator?.hardwareConcurrency > 1;
-    const isWorker = typeof self !== 'undefined' && typeof WorkerGlobalScope !== 'undefined' && self instanceof WorkerGlobalScope;
+    const hasHardwareConcurrency =
+      globalThis.navigator?.hardwareConcurrency > 1;
+    const isWorker =
+      typeof self !== 'undefined' &&
+      typeof WorkerGlobalScope !== 'undefined' &&
+      self instanceof WorkerGlobalScope;
 
     // We only use multi-threading if the browser has threads and we're in a Worker context
     // This is a caveat of threading library we use (wasm-bindgen-rayon)
-    if (isWorker && hasHardwareConcurrency && await threads()) {
+    if (isWorker && hasHardwareConcurrency && (await threads())) {
       wasmReady = initMT(moduleOrPath);
     } else {
       wasmReady = initST(moduleOrPath);
@@ -63,7 +69,11 @@ export default async function optimise(
   options: Partial<OptimiseOptions> = {},
 ): Promise<ArrayBuffer> {
   const _options = { ...defaultOptions, ...options };
-  const optimise = await init();;
-  return optimise(new Uint8Array(data), _options.level, _options.interlace, _options.optimiseAlpha)
-    .buffer;
+  const optimise = await init();
+  return optimise(
+    new Uint8Array(data),
+    _options.level,
+    _options.interlace,
+    _options.optimiseAlpha,
+  ).buffer;
 }