+++ packages/png/decode.ts
@@ -16,12 +16,15 @@
  * and modified it to decode PNG images.
  */
 
-import type { InitInput, InitOutput as PngModule } from './codec/squoosh_png';
-import initPngModule, { decode as pngDecode }  from './codec/squoosh_png';
+import type {
+  InitInput,
+  InitOutput as PngModule,
+} from './codec/squoosh_png.js';
+import initPngModule, { decode as pngDecode } from './codec/squoosh_png.js';
 
 let pngModule: Promise<PngModule>;
 
-export async function init (moduleOrPath?: InitInput): Promise<PngModule> {
+export async function init(moduleOrPath?: InitInput): Promise<PngModule> {
   if (!pngModule) {
     pngModule = initPngModule(moduleOrPath);
   }
@@ -29,9 +32,7 @@ export async function init (moduleOrPath?: InitInput): Promise<PngModule> {
   return pngModule;
 }
 
-export default async function decode(
-  data: ArrayBuffer,
-): Promise<ImageData> {
+export default async function decode(data: ArrayBuffer): Promise<ImageData> {
   await init();
 
   const imageData = await pngDecode(new Uint8Array(data));