+++ packages/oxipng/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## @jsquash/oxipng@2.1.0
+
+### Adds
+
+- Adds Node.js ESM support
+    - Updates relative imports to use file extensions
+    - Adds `module` field to relevant `package.json`
+    - Updates pre.js to polyfill ImageData for Node.js
+
 ## @jsquash/oxipng@2.0.0
 
 ### Breaking Changes